refactor: improve OAuth2 error handling in tests
- Replaces custom RoleException throws with standardized OAuth2AuthenticationException
- Refines exception assertions to include error code checks
- Moves utility functions for OIDC user requests to a separate file for better organization and reuse
