+++ server/src/test/kotlin/org/elaastic/auth/oauth/ElaasticOidcUserServiceIntegrationTest.kt
@@ -17,11 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.mock.mockito.SpyBean
 import org.springframework.context.annotation.Profile
-import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest
-import org.springframework.security.oauth2.client.registration.ClientRegistration
-import org.springframework.security.oauth2.core.AuthorizationGrantType
-import org.springframework.security.oauth2.core.OAuth2AccessToken
-import org.springframework.security.oauth2.core.oidc.OidcIdToken
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException
 import org.springframework.security.oauth2.core.oidc.user.OidcUser
 import javax.transaction.Transactional
 
@@ -226,7 +222,7 @@ class ElaasticOidcUserServiceIntegrationTest(
                 )
             }
         }.tThen("an exception is thrown") {
-            assertThrows<RoleException> {
+            assertOAuth2AuthentificationThrowAndErrorCodeRoleException {
                 it()
             }
         }
@@ -246,7 +242,7 @@ class ElaasticOidcUserServiceIntegrationTest(
                 elaasticOidcUserService.loadUser(getUserRequest(it, "UNKNOWN_ROLE"))
             }
         }.tThen("an exception is thrown") {
-            assertThrows<RoleException> {
+            assertOAuth2AuthentificationThrowAndErrorCodeRoleException {
                 it()
             }
         }
@@ -292,7 +288,7 @@ class ElaasticOidcUserServiceIntegrationTest(
             assertFalse(it.user hasRole anotherRole);
             { elaasticOidcUserService.loadUser(getUserRequest(it.user, anotherRole)) }
         }.tThen("an exception is thrown") {
-            assertThrows<RoleException> {
+            assertOAuth2AuthentificationThrowAndErrorCodeRoleException {
                 it()
             }
         }
@@ -348,66 +344,21 @@ class ElaasticOidcUserServiceIntegrationTest(
         }
     }
 
-    private fun getUserRequest(user: User, role: String): OidcUserRequest = getUserRequest(user, listOf(role))
-
-    private fun getUserRequest(user: User, role: RoleId): OidcUserRequest =
-        getUserRequest(user, listOf(getKeycloakRoleFrom(role)))
-
     /**
-     * Create a UserRequest for the given user and role
+     * Assert that the given block throws an [OAuth2AuthenticationException] with the error code
+     * [RoleException::class.java.simpleName]
      *
-     * To see how the role is added, see `oidcIdToken(User, List<Role.RoleId>)`
-     *
-     * @see oidcIdToken
+     * @param block the block to execute
      */
-    private fun getUserRequest(user: User, roles: List<String>): OidcUserRequest = OidcUserRequest(
-        clientRegistration(user),
-        oAuth2AccessToken(),
-        oidcIdToken(user, roles),
-        emptyMap()
-    )
-
-    private fun oidcIdToken(user: User, roles: List<String>): OidcIdToken {
-        return OidcIdToken(
-            "idToken",
-            null,
-            null,
-            mapOf(
-                "sub" to user.username,
-                "given_name" to user.firstName,
-                "family_name" to user.lastName,
-                "email" to user.email,
-                "iss" to "https://localhost:8080",
-                "realm_access" to mapOf(
-                    "roles" to roles.map(::getKeycloakRoleFrom)
-                ),
-            )
+    private fun assertOAuth2AuthentificationThrowAndErrorCodeRoleException(
+        block: () -> Unit,
+    ) {
+        val exception = assertThrows<OAuth2AuthenticationException> {
+            block()
+        }
+        assertTrue(
+            exception.error.errorCode == RoleException::class.java.simpleName,
+            "The exception should have the error code ${RoleException::class.java.simpleName} but was ${exception.error.errorCode}"
         )
     }
-
-    /** Get the Keycloak role from the given RoleId */
-    fun getKeycloakRoleFrom(role: RoleId): String {
-        return keycloakToElaasticRole.entries.find { it.value == role }?.key
-            ?: throw RoleException("Role $role not found")
-    }
-
-    /**
-     * Get the Keycloak role from the given role name.
-     *
-     * If the role name is not found, it will return the role name itself.
-     */
-    fun getKeycloakRoleFrom(role: String): String {
-        return keycloakToElaasticRole.entries.find { it.value.roleName == role }?.key
-            ?: role
-    }
-
-    private fun oAuth2AccessToken() = OAuth2AccessToken(OAuth2AccessToken.TokenType.BEARER, "accessToken", null, null)
-
-    private fun clientRegistration(user: User): ClientRegistration? =
-        ClientRegistration
-            .withRegistrationId(user.firstName)
-            .authorizationGrantType(AuthorizationGrantType.CLIENT_CREDENTIALS)
-            .clientId(user.firstName)
-            .tokenUri("https://localhost:8080")
-            .build()
 }
